Modify code structure in sidebar front side
