+++ src/sidebar/app/components/editor/data/config.json
@@ -1,0 +1,37 @@
+{
+  "heading": {
+    "options": [
+      {
+        "modelElement": "paragraph",
+        "title": "P",
+        "class": "ck-heading_paragraph"
+      },
+      {
+        "modelElement": "heading3",
+        "viewElement": "h3",
+        "title": "H3",
+        "class": "ck-heading_heading3"
+      },
+      {
+        "modelElement": "heading2",
+        "viewElement": "h2",
+        "title": "H2",
+        "class": "ck-heading_heading2"
+      },
+      {
+        "modelElement": "heading1",
+        "viewElement": "h1",
+        "title": "H1",
+        "class": "ck-heading_heading1"
+      }
+    ]
+  },
+  "toolbar": [
+    "headings",
+    "bold",
+    "italic",
+    "strike",
+    "bulletedList",
+    "numberedList"
+  ]
+}
