+++ src/sidebar/app/constants.js
@@ -1,0 +1,1 @@
+export const SURVEY_PATH = 'https://qsurvey.mozilla.com/s3/notes?ref=sidebar';
