+++ src/sidebar/app/utils.js
@@ -1,0 +1,14 @@
+/**
+ * Formats time for the Notes footer
+ * @param time
+ * @returns {string}
+ */
+function formatFooterTime(date) {
+  date = date || Date.now();
+  return new Date(date).toLocaleTimeString([], {
+    hour: '2-digit',
+    minute: '2-digit'
+  });
+}
+
+export { formatFooterTime };
