+++ src/sidebar/static/css/styles.css
@@ -303,53 +303,6 @@ footer {
   right: 10px;
 }
 
-#sync-note,
-#migration-note {
-  display: none;
-  padding: 3px 0 5px 0;
-  background: #ffe900;
-  width: 100%;
-  position: relative;
-  margin: 0 auto;
-  border: 0;
-  border-top: 1px solid #ccc;
-}
-
-#close-button,
-#migration-close-button {
-  width: 16px;
-  position: absolute;
-  top: -6px;
-  right: 4px;
-  padding: 0 0;
-  opacity: 0.4;
-}
-
-#close-button:hover,
-#migration-close-button:hover {
-  cursor: pointer;
-  opacity: 0.6;
-}
-
-#close-button:active,
-#migration-close-button:active {
-  cursor: pointer;
-  opacity: 0.8;
-}
-
-#sync-note-dialog,
-#migration-note-dialog {
-  width: 85%;
-  font-size: 12px;
-  margin: 5px 0 5px 10px;
-  color: #424242;
-}
-
-#sync-note.visible,
-#migration-note.visible {
-  display: block;
-}
-
 /* Select menu from sync */
 #context-menu-button {
   background-size: 3px;
